requirements.txt
numpy
pandas
scipy
scikit-learn
Pillow
matplotlib
